+++ app/views/families/show.html.erb
